Added try/catch to DeepDive method.
1) Possible exceptions for GetCurrentDirectory method:

UnauthorizedAccessException
The caller does not have the required permission.

NotSupportedException
The operating system is Windows CE, which does not have current directory functionality.

2) Possible exceptions for CreateDirectory method:

IOException
The directory specified by path is a file.

UnauthorizedAccessException
The caller does not have the required permission.

ArgumentException
path is a zero-length string, contains only white space, or contains one or more invalid characters. You can query for invalid characters by using the GetInvalidPathChars() method.

-or-

path is prefixed with, or contains, only a colon character (:).

ArgumentNullException
path is null.

PathTooLongException
The specified path, file name, or both exceed the system-defined maximum length.

DirectoryNotFoundException
The specified path is invalid (for example, it is on an unmapped drive).

NotSupportedException
path contains a colon character (:) that is not part of a drive label ("C:\").